+++ src/main/java/com/example/demo/domain/enums/Rating.java
@@ -9,24 +9,24 @@ import java.math.BigDecimal;
 public enum Rating {
 
 	LOW {
-		public double calculate(BigDecimal totalRating, int count) {
-			return new LowSatisfactionAverageCalculator(totalRating.doubleValue(), count).calculateAverage();
+		public BigDecimal calculate(BigDecimal totalRating, int count) {
+			return new LowSatisfactionAverageCalculator(totalRating, count).calculateAverage();
 		}
 	},
 
 	DECENT {
-		public double calculate(BigDecimal totalRating, int count) {
-			return new DecentSatisfactionAverageCalculator(totalRating.doubleValue(), count).calculateAverage();
+		public BigDecimal calculate(BigDecimal totalRating, int count) {
+			return new DecentSatisfactionAverageCalculator(totalRating, count).calculateAverage();
 		}
 	},
 
 	HIGH {
-		public double calculate(BigDecimal totalRating, int count) {
-			return new HighSatisfactionAverageCalculator(totalRating.doubleValue(), count).calculateAverage();
+		public BigDecimal calculate(BigDecimal totalRating, int count) {
+			return new HighSatisfactionAverageCalculator(totalRating, count).calculateAverage();
 		}
 	};
 
-	public abstract double calculate(BigDecimal totalRating, int count);
+	public abstract BigDecimal calculate(BigDecimal totalRating, int count);
 
 	public static Rating mapToSatisfactionRate(Double rating) {
 		if (rating < 5D || rating == null) {
